rename package to DQMOffline/Lumi, rename namespace, use unique_ptr
