+++ DQMOffline/Lumi/interface/MiniBaconDefs.h
@@ -1,5 +1,5 @@
-#ifndef DQMOFFLINE_LUMIZCOUNTING_MINIBACONDEFS_H
-#define DQMOFFLINE_LUMIZCOUNTING_MINIBACONDEFS_H
+#ifndef DQMOFFLINE_LUMI_MINIBACONDEFS_H
+#define DQMOFFLINE_LUMI_MINIBACONDEFS_H
 
 #include <bitset>
