+++ DQMOffline/Lumi/plugins/BuildFile.xml
@@ -5,7 +5,7 @@
 <use name="DQMServices/Examples"/>
 <use name="DQMServices/Core"/>
 <use name="DataFormats/CLHEP"/>
-<use name="DQMOffline/LumiZCounting"/>
+<use name="DQMOffline/Lumi"/>
 <library file="*.cc" name="ZCountingPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
